add role manager to remove hotel

+++ src/main/java/com/hust/smarthotel/components/mananging/controller/ManagingHotelController.java
@@ -5,11 +5,11 @@ import com.hust.smarthotel.components.hotel.domain_service.HotelService;
 import com.hust.smarthotel.components.mananging.app_model.HotelRequest;
 import com.hust.smarthotel.components.mananging.app_model.ManagingResponse;
 import com.hust.smarthotel.components.mananging.domain_model.ExtendedManaging;
+import com.hust.smarthotel.components.mananging.domain_model.Managing;
 import com.hust.smarthotel.components.mananging.domain_service.ManagingService;
 import com.hust.smarthotel.components.user.domain_model.User;
 import com.hust.smarthotel.components.user.domain_service.UserService;
 import com.hust.smarthotel.generic.constant.HeaderConstant;
-import com.hust.smarthotel.generic.constant.RoleConstants;
 import com.hust.smarthotel.generic.constant.UrlConstants;
 import com.hust.smarthotel.generic.util.JwtUtil;
 import io.jsonwebtoken.Claims;
@@ -25,9 +25,9 @@ import javax.validation.Valid;
 import static com.hust.smarthotel.generic.response.ErrorResponses.MANAGING_USER_NOT_FOUND;
 import static com.hust.smarthotel.generic.response.ErrorResponses.MANAGING_NOT_MANAGER;
 import static com.hust.smarthotel.generic.response.ErrorResponses.MANAGING_HOTEL_NOT_FOUND;
+import static com.hust.smarthotel.generic.response.ErrorResponses.MANAGING_INVALID_TOKEN;
 
 import static com.hust.smarthotel.generic.constant.RoleConstants.MANAGER;
-import static com.hust.smarthotel.generic.constant.RoleConstants.ADMIN;
 
 
 @RestController
@@ -38,6 +38,8 @@ public class ManagingHotelController {
     private static final ResponseEntity<ManagingResponse> USER_NOT_FOUND = new ResponseEntity<>(MANAGING_USER_NOT_FOUND, HttpStatus.BAD_REQUEST);
     private static final ResponseEntity<ManagingResponse> NOT_MANAGER = new ResponseEntity<>(MANAGING_NOT_MANAGER, HttpStatus.BAD_REQUEST);
     private static final ResponseEntity<ManagingResponse> HOTEL_NOT_FOUND = new ResponseEntity<>(MANAGING_HOTEL_NOT_FOUND, HttpStatus.BAD_REQUEST);
+    private static final ResponseEntity<ManagingResponse> INVALID_TOKEN = new ResponseEntity<>(MANAGING_INVALID_TOKEN, HttpStatus.FORBIDDEN);
+
 
     @Autowired
     private UserService userService;
@@ -92,10 +94,17 @@ public class ManagingHotelController {
     }
 
     @DeleteMapping("/{managerId}/hotels/{hotelId}")
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    ResponseEntity<ManagingResponse> removeHotelFromManager(@PathVariable("managerId") String managerId,
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_MANAGER')")
+    ResponseEntity<ManagingResponse> removeHotelFromManager(@RequestHeader(value = HeaderConstant.AUTHORIZATION) String authorizationField,
+                                                            @PathVariable("managerId") String managerId,
                                                             @PathVariable("hotelId") String hotelId){
-        System.out.println("Remove hotel " + hotelId + " from manager " + managerId);
+        String token = authorizationField.replace(HeaderConstant.TOKEN_PREFIX, "");
+        Claims claims = jwtUtil.getClaims(token);
+        String userId = claims.getSubject();
+        String role = claims.get(JwtUtil.ROLE, String.class);
+        if (role.equals(MANAGER) && !userId.equals(managerId))
+            return INVALID_TOKEN;
+
         ManagingResponse managingResponse = managingService.removeHotelFromManager(managerId, hotelId);
         if (!managingResponse.getStatus())
             return new ResponseEntity<>(managingResponse, HttpStatus.BAD_REQUEST);
